+++ src/main/kotlin/lotto/domain/RandomLottoGenerator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class RandomLottoGenerator: LottoGenerator {
