+++ src/test/kotlin/lotto/domain/DummyLottoGenerator.kt
@@ -1,5 +1,7 @@
-package lotto
+package lotto.domain
 
+import lotto.domain.Lotto
+import lotto.domain.LottoGenerator
 import java.lang.RuntimeException
 
 class DummyLottoGenerator: LottoGenerator {
