+++ src/test/kotlin/lotto/domain/LottoNumberTest.kt
@@ -1,5 +1,6 @@
-package lotto
+package lotto.domain
 
+import lotto.domain.LottoNumber
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
