+++ src/test/kotlin/lotto/domain/RandomLottoGeneratorTest.kt
@@ -1,5 +1,6 @@
-package lotto
+package lotto.domain
 
+import lotto.domain.RandomLottoGenerator
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
